refactor: remove two module
