+++ pyproject.toml
@@ -101,8 +101,7 @@ dependencies = [
 ]
 exclude = [
     "src/psygnal/__init__.py",
-    "src/psygnal/_evented_model_v1.py",
-    "src/psygnal/_evented_model_v2.py",
+    "src/psygnal/_evented_model_shared.py",
     "src/psygnal/utils.py",
     "src/psygnal/containers",
     "src/psygnal/qt.py",
