+++ src/psygnal/__init__.py
@@ -8,7 +8,7 @@ from importlib.metadata import PackageNotFoundError, version
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from ._evented_model_v1 import EventedModel  # noqa: TCH004
+    from ._evented_model_shared import EventedModel  # noqa: TCH004
 
 
 try:
@@ -63,12 +63,7 @@ from ._throttler import debounced, throttled
 
 def __getattr__(name: str) -> Any:
     if name == "EventedModel":
-        import pydantic.version
-
-        if pydantic.version.VERSION.startswith("2"):
-            from ._evented_model_v2 import EventedModel
-        else:
-            from ._evented_model_v1 import EventedModel  # type: ignore
+        from ._evented_model_shared import EventedModel
 
         return EventedModel
     raise AttributeError(  # pragma: no cover
